requirements.txt
python-decouple==3.8
python-telegram-bot==13.2
requests==2.32.5